Fix Reader State Restoration
This fix addresses an issue where  would display a loading screen when returning to the app after the process was killed by the system. By proactively storing the manga ID and reloading the data in the ViewModel, the app can now seamlessly restore the reader state without any noticeable lag.